TASK-29 minor refactor
