+++ src/app.tsx
@@ -1,6 +1,8 @@
 import React from "react";
+import { Route, Switch } from "react-router-dom";
 import { CustomGlobalHotKeys, HeaderComponent } from "./components/common-components";
-import ViewOnlyComponent from "./components/schedule-page/view-only.component";
+import { ViewOnlyComponent } from "./components/view-only.component";
+import { ScheduleEditingComponent } from "./components/schedule-page/schedule-editing.component";
 
 function App(): JSX.Element {
   return (
@@ -8,7 +10,10 @@ function App(): JSX.Element {
       <div>
         <CustomGlobalHotKeys />
         <HeaderComponent />
-        <ViewOnlyComponent />
+        <Switch>
+          <Route path="/" component={ViewOnlyComponent} exact />
+          <Route path="/schedule-editing" component={ScheduleEditingComponent} />
+        </Switch>
       </div>
     </React.Fragment>
   );
