+++ src/components/schedule-page/schedule-page.component.tsx
@@ -1,0 +1,12 @@
+import React from "react";
+import { TableComponent } from "./table/table.component";
+
+export default function SchedulePage(): JSX.Element {
+  return (
+    <React.Fragment>
+      <div className="cols-3-to-1">
+        <TableComponent />
+      </div>
+    </React.Fragment>
+  );
+}
